Prettier

+++ contracts/RemoteAddressValidator.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity 0.8.9;    
+pragma solidity 0.8.9;
 import { IRemoteAddressValidator } from './interfaces/IRemoteAddressValidator.sol';
 import { StringToAddress, AddressToString } from '@axelar-network/axelar-gmp-sdk-solidity/contracts/StringAddressUtils.sol';
 import { Upgradable } from '@axelar-network/axelar-gmp-sdk-solidity/contracts/upgradables/Upgradable.sol';
@@ -19,27 +19,31 @@ contract RemoteAddressValidator is IRemoteAddressValidator, Upgradable {
     // bytes32(uint256(keccak256('remote-address-validator')) - 1)
     //bytes32 public constant override contractId = 0x5d9f4d5e6bb737c289f92f2a319c66ba484357595194acb7c2122e48550eda7c;
 
-    constructor (address tokenLinkerAddress_, string[] memory trustedChainNames, string[] memory trustedAddresses ) {
-        if(tokenLinkerAddress_ == address(0)) revert ZeroAddress();
+    constructor(
+        address tokenLinkerAddress_,
+        string[] memory trustedChainNames,
+        string[] memory trustedAddresses
+    ) {
+        if (tokenLinkerAddress_ == address(0)) revert ZeroAddress();
         tokenLinkerAddress = tokenLinkerAddress_;
         uint256 length = trustedChainNames.length;
-        if(length != trustedAddresses.length) revert LengthMismatch();
+        if (length != trustedAddresses.length) revert LengthMismatch();
         addressHash = keccak256(bytes(tokenLinkerAddress.toString()));
-        for(uint i; i<length; ++i) {
+        for (uint256 i; i < length; ++i) {
             string memory chainName = trustedChainNames[i];
-            if(bytes(chainName).length == 0) revert ZeroStringLength();
+            if (bytes(chainName).length == 0) revert ZeroStringLength();
             string memory remoteAddress = trustedAddresses[i];
-            if(bytes(remoteAddress).length == 0) revert ZeroStringLength();
+            if (bytes(remoteAddress).length == 0) revert ZeroStringLength();
             remoteAddresses[chainName] = remoteAddress;
             bytes32 remoteAddressHash = keccak256(bytes(remoteAddress));
             remoteAddressHashes[chainName] = remoteAddressHash;
         }
     }
 
-    function _lowerCase(string memory s) internal pure returns(string memory) {
+    function _lowerCase(string memory s) internal pure returns (string memory) {
         uint256 length = bytes(s).length;
         bytes memory tmp = bytes(s);
-        for(uint256 i; i<length; i++) {
+        for (uint256 i; i < length; i++) {
             uint8 b = uint8(tmp[i]);
             if ((b >= 65) && (b <= 70)) tmp[i] = bytes1(b + uint8(32));
         }
@@ -49,16 +53,16 @@ contract RemoteAddressValidator is IRemoteAddressValidator, Upgradable {
     function validateSender(string calldata sourceChain, string calldata sourceAddress) external view override returns (bool) {
         string memory sourceAddressLC = _lowerCase(sourceAddress);
         bytes32 sourceAddressHash = keccak256(bytes(sourceAddressLC));
-        if(sourceAddressHash == addressHash) return true;
-        if(sourceAddressHash == remoteAddressHashes[sourceChain]) return true;
+        if (sourceAddressHash == addressHash) return true;
+        if (sourceAddressHash == remoteAddressHashes[sourceChain]) return true;
         return false;
     }
-
 
     function addTrustedAddress(string calldata sourceChain, string calldata sourceAddress) external onlyOwner {
         remoteAddressHashes[sourceChain] = keccak256(bytes(sourceAddress));
         remoteAddresses[sourceChain] = sourceAddress;
     }
+
     function removeTrustedAddress(string calldata sourceChain) external onlyOwner {
         remoteAddressHashes[sourceChain] = bytes32(0);
         remoteAddresses[sourceChain] = '';
@@ -66,7 +70,7 @@ contract RemoteAddressValidator is IRemoteAddressValidator, Upgradable {
 
     function getRemoteAddress(string calldata chainName) external view override returns (string memory remoteAddress) {
         remoteAddress = remoteAddresses[chainName];
-        if(bytes(remoteAddress).length == 0) {
+        if (bytes(remoteAddress).length == 0) {
             remoteAddress = tokenLinkerAddress.toString();
         }
     }
